+++ examples/general/plot_12_group_glm.py
@@ -167,7 +167,9 @@ def individual_analysis(bids_path, ID):
     cha = glm_est.to_dataframe()
 
     # Compute region of interest results from channel data
-    roi = glm_est.to_dataframe_region_of_interest(groups, design_matrix.columns)
+    roi = glm_est.to_dataframe_region_of_interest(groups,
+                                                  design_matrix.columns,
+                                                  demographic_info=True)
 
     # Define left vs right tapping contrast
     contrast_matrix = np.eye(design_matrix.shape[1])
@@ -282,6 +284,35 @@ roi_model.summary()
 
 
 ###############################################################################
+# Second level analysis with covariates
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+#
+# .. sidebar:: Relevant literature
+#
+#    For a detailed discussion about covariates in fNIRS analysis see
+#    the seminar by Dr. Jessica Gemignani
+#    (`youtube <https://www.youtube.com/watch?feature=emb_logo&v=3E28sT1JI14>`_).
+#
+# It is simple to extend these models to include covariates.
+# This dataset is small, so including additional factors may not be
+# appropriate. However, for instructional purpose, we will include a
+# covariate of gender. Also, for instructional purpose, we modify the model
+# above to only explore the difference between the two tapping conditions in
+# the hbo signal in the right hemisphere.
+#
+# From the model result we observe that hbo responses in the right hemisphere
+# are smaller when the right hand was used (as expected for these
+# contralaterally dominant responses) and there is no effect of gender.
+
+grp_results = df_roi.query("Condition in ['Tapping/Left', 'Tapping/Right']")
+grp_results = grp_results.query("Chroma in ['hbo']")
+grp_results = grp_results.query("ROI in ['Right_Hemisphere']")
+
+roi_model = smf.mixedlm("theta ~ Condition + Sex",
+                        grp_results, groups=grp_results["ID"]).fit(method='nm')
+roi_model.summary()
+
+###############################################################################
 # Visualise group results
 # -----------------------
 #
@@ -293,6 +324,11 @@ roi_model.summary()
 # We also observe the the tapping response is
 # larger in the contralateral hemisphere.
 # Filled symbols represent HbO, unfilled symbols represent HbR.
+
+# Regenerate the results from the original group model above
+grp_results = df_roi.query("Condition in ['Control','Tapping/Left', 'Tapping/Right']")
+roi_model = smf.mixedlm("theta ~ -1 + ROI:Condition:Chroma",
+                        grp_results, groups=grp_results["ID"]).fit(method='nm')
 
 df = statsmodels_to_results(roi_model)